requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
PyYAML==6.0.1
reportlab==4.0.7
python-dotenv==1.0.0
openai==1.3.7
schedule==1.2.0
selenium==4.15.2
webdriver-manager==4.0.1
flask==2.3.3
